requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-cors==4.0.0
pymysql==1.1.0
cryptography==41.0.3
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
werkzeug==2.3.7